Trait pallet_liquidity_pools_gateway::pallet::Config

pub trait Config: Config {
    type RuntimeOrigin: Into<Result<GatewayOrigin, <Self as Config>::RuntimeOrigin>> + From<GatewayOrigin>;
    type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>;
    type LocalEVMOrigin: EnsureOrigin<<Self as Config>::RuntimeOrigin, Success = GatewayOrigin>;
    type AdminOrigin: EnsureOrigin<<Self as Config>::RuntimeOrigin>;
    type Message: Codec + Clone + Debug + PartialEq;
    type Router: DomainRouter<Sender = Self::AccountId, Message = Self::Message> + Clone + Debug + MaxEncodedLen + TypeInfo + FullCodec + EncodeLike + PartialEq;
    type InboundQueue: InboundQueue<Sender = DomainAddress, Message = Self::Message>;
    type OriginRecovery: TryConvert<(Vec<u8>, Vec<u8>), DomainAddress, Error = DispatchError>;
    type WeightInfo: WeightInfo;
    type MaxIncomingMessageSize: Get<u32>;
    type Sender: Get<Self::AccountId>;
}

Configuration trait of this pallet.

Implement this type for a runtime in order to customize this pallet.

Required Associated Types

type RuntimeOrigin: Into<Result<GatewayOrigin, <Self as Config>::RuntimeOrigin>> + From<GatewayOrigin>

The origin type.

type RuntimeEvent: From<Event<Self>> + IsType<<Self as Config>::RuntimeEvent>

The event type.

type LocalEVMOrigin: EnsureOrigin<<Self as Config>::RuntimeOrigin, Success = GatewayOrigin>

The LocalOrigin ensures that some calls can only be performed from a local context i.e. a different pallet.

type AdminOrigin: EnsureOrigin<<Self as Config>::RuntimeOrigin>

The AdminOrigin ensures that some calls can only be performed by admins.

type Message: Codec + Clone + Debug + PartialEq

The incoming and outgoing message type.

NOTE - this Codec trait is the Centrifuge trait for liquidity pools’ messages.

type Router: DomainRouter<Sender = Self::AccountId, Message = Self::Message> + Clone + Debug + MaxEncodedLen + TypeInfo + FullCodec + EncodeLike + PartialEq

The message router type that is stored for each domain.

type InboundQueue: InboundQueue<Sender = DomainAddress, Message = Self::Message>

The type that processes incoming messages.

type OriginRecovery: TryConvert<(Vec<u8>, Vec<u8>), DomainAddress, Error = DispatchError>

A way to recover a domain address from two byte slices

type WeightInfo: WeightInfo

type MaxIncomingMessageSize: Get<u32>

Maximum size of an incoming message.

type Sender: Get<Self::AccountId>

The sender account that will be used in the OutboundQueue implementation.